Safari-611.1.10/LayoutTests/fast/events/touch/gesture/gesture-scrollbar-expected.txt

This tests scroll gesture events on a scroll bar. The scrollable div below should be slightly scrolled down if successful.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS movingDiv.scrollTop is 0
PASS movingDiv.scrollTop is >= 1
PASS successfullyParsed is true

TEST COMPLETE
